test: stabilize browser regression harness
