+++ apps/desktop/tooling/vite.config.ts
@@ -3,100 +3,12 @@ import react from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
 import { readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
-import type { BrowserCommandContext } from "vitest/node";
 import { configDefaults, defineConfig } from "vitest/config";
 import { z } from "zod";
 
 const protocolVersionDefinition = z
   .object({ version: z.string().min(1) })
   .parse(JSON.parse(readFileSync(new URL("../../../shared/protocol/version.json", import.meta.url), "utf8")));
-
-export type ReorderPointerCommandInput = Readonly<{
-  sourceSelector: string;
-  destination:
-    | Readonly<{ kind: "source" }>
-    | Readonly<{
-        kind: "target";
-        placement: "center" | "gap";
-        selector: string;
-      }>;
-}>;
-
-declare module "vitest/internal/browser" {
-  interface BrowserCommands {
-    pointerDrag: (input: ReorderPointerCommandInput) => Promise<void>;
-  }
-}
-
-type PointerBox = Readonly<{
-  height: number;
-  width: number;
-  x: number;
-  y: number;
-}>;
-
-type PointerFrame = Readonly<{
-  locator(selector: string): Readonly<{
-    boundingBox(): Promise<PointerBox | null>;
-  }>;
-}>;
-
-type PointerCommandRuntimeContext = Readonly<{
-  frame(): Promise<PointerFrame>;
-  page: Readonly<{
-    mouse: Readonly<{
-      down(): Promise<void>;
-      move(x: number, y: number): Promise<void>;
-      up(): Promise<void>;
-    }>;
-  }>;
-}>;
-
-const pointerCommandRuntimeContextSchema = z.custom<PointerCommandRuntimeContext>(
-  (value) => value instanceof Object,
-);
-
-export const pointerDrag = async (
-  { provider, sessionId }: BrowserCommandContext,
-  input: ReorderPointerCommandInput,
-): Promise<void> => {
-  const runtimeContext = pointerCommandRuntimeContextSchema.parse(provider.getCommandsContext(sessionId));
-  const testFrame = await runtimeContext.frame();
-  const sourceBox = await resolveBox(testFrame, input.sourceSelector, "source");
-  const destinationBox =
-    input.destination.kind === "source"
-      ? sourceBox
-      : await resolveBox(testFrame, input.destination.selector, "destination");
-  const destination = projectPointerDestination(destinationBox, input.destination);
-  const sourceX = sourceBox.x + sourceBox.width / 2;
-  const sourceY = sourceBox.y + sourceBox.height / 2;
-  await runtimeContext.page.mouse.move(sourceX, sourceY);
-  await runtimeContext.page.mouse.down();
-  await runtimeContext.page.mouse.move(sourceX, sourceY + 7);
-  await runtimeContext.page.mouse.move(destination.x, destination.y);
-  await runtimeContext.page.mouse.up();
-};
-
-async function resolveBox(frame: PointerFrame, selector: string, role: string): Promise<PointerBox> {
-  const box = await frame.locator(selector).boundingBox();
-  if (box === null) {
-    throw new Error(`Cannot locate pointer-drag ${role}: ${selector}`);
-  }
-  return box;
-}
-
-function projectPointerDestination(
-  box: PointerBox,
-  destination: ReorderPointerCommandInput["destination"],
-): Readonly<{ x: number; y: number }> {
-  if (destination.kind === "source") {
-    return { x: box.x + box.width / 2, y: box.y + box.height / 2 };
-  }
-  return {
-    x: box.x + box.width / 2,
-    y: destination.placement === "center" ? box.y + box.height / 2 : box.y - 4,
-  };
-}
 
 export default defineConfig({
   plugins: [tailwindcss(), react()],
