+++ apps/desktop/tooling/vitest.browser.config.mjs
@@ -1,12 +1,51 @@
 import { env } from "node:process";
 import { mergeConfig, defineConfig } from "vitest/config";
 import { playwright } from "@vitest/browser-playwright";
-import baseConfig, { pointerDrag } from "./vite.config";
+import baseConfig from "./vite.config";
+
+const pointerDrag = async ({ provider, sessionId }, input) => {
+  const runtimeContext = provider.getCommandsContext(sessionId);
+  if (!(runtimeContext instanceof Object)) {
+    throw new Error("Cannot resolve the browser pointer-drag runtime context.");
+  }
+  const testFrame = await runtimeContext.frame();
+  const sourceBox = await resolveBox(testFrame, input.sourceSelector, "source");
+  const destinationBox =
+    input.destination.kind === "source"
+      ? sourceBox
+      : await resolveBox(testFrame, input.destination.selector, "destination");
+  const destination = projectPointerDestination(destinationBox, input.destination);
+  const sourceX = sourceBox.x + sourceBox.width / 2;
+  const sourceY = sourceBox.y + sourceBox.height / 2;
+  await runtimeContext.page.mouse.move(sourceX, sourceY);
+  await runtimeContext.page.mouse.down();
+  await runtimeContext.page.mouse.move(sourceX, sourceY + 7);
+  await runtimeContext.page.mouse.move(destination.x, destination.y);
+  await runtimeContext.page.mouse.up();
+};
+
+async function resolveBox(frame, selector, role) {
+  const box = await frame.locator(selector).boundingBox();
+  if (box === null) {
+    throw new Error(`Cannot locate pointer-drag ${role}: ${selector}`);
+  }
+  return box;
+}
+
+function projectPointerDestination(box, destination) {
+  if (destination.kind === "source") {
+    return { x: box.x + box.width / 2, y: box.y + box.height / 2 };
+  }
+  return {
+    x: box.x + box.width / 2,
+    y: destination.placement === "center" ? box.y + box.height / 2 : box.y - 4,
+  };
+}
 
 const chromiumExecutablePath = env.KENT_PLAYWRIGHT_CHROMIUM_PATH;
 const browserProvider =
   chromiumExecutablePath === undefined
-    ? playwright()
+    ? playwright({ launchOptions: { channel: "chromium" } })
     : playwright({ launchOptions: { executablePath: chromiumExecutablePath } });
 
 const browserConfig = mergeConfig(
